+++ R/DAISIE_loglik_IW.R
@@ -1,28 +1,3 @@
-#' IW concurrency control
-#'
-#' Sets or retrieves the number of threads used by the odeint solver.
-#'
-#' @param num_threads \code{num_threads < 0 or omitted}: retrieves number of threads. \cr
-#' \code{num_threads = 0}: sets the number of threads to the number of available cores. \cr
-#' \code{num_threads = 1}: single-threaded execution. \cr
-#' \code{num_threads > 1}: sets the number of threads to \code{num_threads}.
-#' @return number of threads
-#' @note The maximum number of threads is limited to the value of the C++
-#' standard library function \code{std::thread::hardware_concurrency()}.
-#' This is also the default number of threads upon library load.
-#' Multithreading incurs some overhead. Therefore, single-threaded execution
-#' might be faster for small systems.
-#'
-#' @export DAISIE_IW_num_threads
-DAISIE_IW_num_threads <- function(num_threads) {
-  if (missing(num_threads)) {
-    # retrieve only
-    return(.Call("daisie_odeint_iw_num_threads", -1))
-  }
-  return(.Call("daisie_odeint_iw_num_threads", num_threads))
-}
-
-
 dec2bin <- function(y, ly) {
   stopifnot(length(y) == 1, mode(y) == "numeric")
   q1 <- (y / 2) %/% 1
@@ -399,7 +374,7 @@ DAISIE_loglik_IW <- function(
     parslist <- list(pars = pars1,k = k,ddep = ddep,dime = dime,l0ki = l0ki,nndd = nndd)
     iw_parms = DAISIE_IW_pars(parslist)
     if (startsWith(methode, "odeint::")) {
-      probs <- .Call("daisie_odeint_iw", probs, brts[(k + 1):(k + 2)], iw_parms, methode, abstolint, reltolint)
+      probs <- Daisie_odeint_iw(probs, brts[(k + 1):(k + 2)], iw_parms, methode, abstolint, reltolint)
     } else {
       y <- deSolve::ode(y = probs,
                         times = brts[(k + 1):(k + 2)],
@@ -499,7 +474,7 @@ DAISIE_loglik_IW <- function(
     parslist <- list(pars = pars1,k = 0,ddep = ddep,dime = dime,l0ki = l0ki,nndd = nndd)
     iw_parms <- DAISIE_IW_pars(parslist)
     if (startsWith(methode, "odeint::")) {
-      probs <- .Call("daisie_odeint_iw", probs, c(min(brts),0), iw_parms, methode, abstolint, reltolint)
+      probs <- DAISIE_odeint_iw(probs, c(min(brts),0), iw_parms, methode, abstolint, reltolint)
     } else {
       y <- deSolve::ode(y = probs,
                         times = c(min(brts), 0),